Add Connector to Infoblox WAPI
Adding a bit refactored connector that was used for a while in
infoblox-designate and other neutron related projects.
Improved test coverage.
Previously Infoblox exceptions dependent on neutron exceptions.
But it would be inefficient to depend on huge neutron just to use
the same exception model. So slightly modified exceptions to make
it independent from neutron.