Allow passing options as dict for connector
Previously connector expected config object that with appropriate
fields.
And config is not a part of this package, so need to add flexibility in
the way how options should be passed.
Since connector is now in another package to config (it stays in
networking-infoblox) more flexible way of passing parameters has to be
introduced.
This changes allows to pass options as a dict.