Add new field type '_updateable_search_field'
Previously if field was marked as search field it was treated as
unchangeable, so it was not sent on objects update.
But some fields can act as both search field and updateable field,
for example 'name' in HostRecord.
So new type of fields is introduced '_updateable_search_field'.
Adding field to this group changes update behaviour only and
field is always sent on update to NIOS idependently from search_fields
values.

Updated HostRecordV4 and HostRecordV6 to treat 'name' field as
updateable search field.

Closes: #21