Raise exception with details if reply is not json
Previously if NIOS replies in not json format ValueError was raised from
json library, but reply itself was not logged.
Typically if reply is not in json format it indicates some error
message.
Updated parsing of replies to raise an exception with reply details if
reply is not json format.

Closes: #25