Add 'silent_ssl_warnings' option to connector
If 'ssl_verify' is set to False, then each api request generates ssl
warnings.
Added new option for connector to be able to disable these warnings.
By default this option is turned off.
This option is not expected to be used in production, as well as
'ssl_verify' = False, so make sure 'silent_ssl_warnings' usage is
limited with development/test environments.

Note: warnings are disabled on urllib3 level, so it affects all
instanses of requests as a result. Use with option cautiously.

Closes: #22