Remove network_exists and change get_network
Network_exists method was quite ineffective way of checking network
existence. It does not return network, so if any action has to
be performed on network then new API call has to be done to get network.
Using get_network instead is more effective, it returns None is no
network was found.

Also get_network was updated to not throw exception if no network is
found, so just None is returned.

Closes: #37