Convert strings into booleans for ssl_verify
ssl_verify for Connector was passed to requests.Session.verify without
changes. ssl_verify can be boolean or string.
If it is a string it was excected to contain path to certificate.

To simplify using ssl_verify option now it can be passed in as string,
that containg anything that look like True or False, i.e. 'true', 'on',
'yes', 'ON' or 'FALSE', 'no', 'off' (all values are case insensitive).
In this case string would be converted to boolean and passed into requests.
All other values are passed as is, i.e. booleans are passed unchanged,
as well as any string that doesn't look like a boolean.

Added try_value_to_bool it has two modes of converitng to booleans.
For EA processing strict mode is needed.
In this mode only 'True' and 'False' string are converted into booleans,
and this check is case sensitive.
In not strict mode anything that means True or False is converted to
booleans, this check case insensitive.
'true/on/yes/false/off/no' are processed in not strict mode.

Closes: #65