Add max_results option to connector and objects
max_results option is translated into '_max_results' WAPI option:

Maximum number of objects to be returned. If set to a negative
number the appliance will return an error when the number of returned
objects would exceed the setting. The default is -1000. If this is set
to a positive number, the results will be truncated when necessary.

This option is available for get_object method in connector and
for search and search_all methods for objects that are inherited from
InfobloxObject.

Closes: #62