Do not override verify flag on request level
Previously 'verify' flag was set on two levels: session and request.
But since it is expected to have the same value in both cases there is
no need to do so.
'verify' flag set on request level has not converted string to booleans.
It caused failures with networking-infoblox usage.

Removed 'verify' flag from request level, so now 'verify' value set on
session level is used.

Closes: #83