Add 'max_results' as connector option
'max_results' defines maximum number of objects to be returned.
If set to a negative number the appliance will return an error
when the number of returned objects would exceed the setting.
The default is -1000. If this is set to a positive number,
the results will be truncated when necessary.

'max_results' was option for get_object method for a while.
Adding it as connector option to set it once and globally.
If 'max_results' defined as connector option, then it applied to
all get_object calls.
'max_results' value can be redefined for particular get_object calls
by passing another 'max_results' value to get_object.

Closes: #101